Optimizing calculation of MediumUnknownWordProb.
Realizing that when all words of a pattern-set are tested for each pattern in the set, there is no need to test each and every pattern (as they all rely on the same set of words). Except that expressions with wildcards needs to be handled individually to match the wildcards.